Fix senator quests generating with 0 threat points
Looks like the senator quests didn't specify amount of threat points to generate with, thus ending up in the default value of 0 being used for them all. This resulted in quests basically always generating with a single enemy defending the quest.

I've changed the quest generation to include points. I've used `StorytellerUtility.DefaultSiteThreatPointsNow()`, as it uses the world threat points and scales them to appropriate site threat points.

On top of that, I've replaced the use of `asker` string in the slate with using the constant of `SlateNames.Asker`. This should ensure that it's always correct, nobody accidentally modifies it, or any similar situations like those.